rm deprecated stuff
