+++ autd3-driver/src/firmware/operation/mod.rs
@@ -5,7 +5,6 @@ mod force_fan;
 mod gain;
 mod gpio_in;
 mod info;
-mod info2;
 mod modulation;
 mod null;
 mod pulse_width_encoder;
@@ -22,7 +21,6 @@ pub use force_fan::*;
 pub use gain::*;
 pub use gpio_in::*;
 pub use info::*;
-pub use info2::*;
 pub use modulation::*;
 pub use null::*;
 pub use pulse_width_encoder::*;
